base/modules/utils/requirements.txt
bson
cbor2
fastavro
kombu
msgpack
xmltodict
pyrabbit
pyyaml
py-ubjson
sexpdata